Clippy + Actions

+++ src/std.rs
@@ -86,7 +86,7 @@ impl Add<Duration> for Instant {
     fn add(self, other: Duration) -> Instant {
         let new_val = self.inner + other.as_millis() as f64;
         Instant {
-            inner: new_val as f64,
+            inner: new_val,
         }
     }
 }
@@ -103,7 +103,7 @@ impl Sub<Duration> for Instant {
     fn sub(self, other: Duration) -> Instant {
         let new_val = self.inner - other.as_millis() as f64;
         Instant {
-            inner: new_val as f64,
+            inner: new_val,
         }
     }
 }
@@ -197,7 +197,7 @@ impl Add<Duration> for SystemTime {
     fn add(self, other: Duration) -> SystemTime {
         let new_val = self.inner + other.as_millis() as f64;
         SystemTime {
-            inner: new_val as f64,
+            inner: new_val,
         }
     }
 }
@@ -208,7 +208,7 @@ impl Sub<Duration> for SystemTime {
     fn sub(self, other: Duration) -> SystemTime {
         let new_val = self.inner - other.as_millis() as f64;
         SystemTime {
-            inner: new_val as f64,
+            inner: new_val,
         }
     }
 }
